am62x/am62ax/am62px: 09.01.00.39 release
Common:
    - Boottime optimization: Use Authentication type 0
    - Boottime optimization: Use BCDMA for DDR to DDR copy during SBL

AM62X:
    - SK-AM62-SIP support for all examples

AM62A:

    - WKUP R5F:
        - Priming DDR for inline ECC done through BIST
        - SBL SD Linux Support
    - MCU R5F:
        - Example for isolation of MCU through firewall
    - A53:
        - FreeRTOS-SMP quad core support
        - ECAP, EEPROM, EPWM, MCASP, MCSPI, MMCSD, OSPI, WDT, SA3UL
    - C7X:
        - GPIO

AM62P:
    - WKUP R5F:
        - SBL: OSPI NOR, EMMC, UART, NULL
        - UART, SCI Client, IPC, GPIO, I2C, MCAN, MCSPI, DDR, OSPI, EMMC
        - DSS Support with freeze frame detect, data integrity check
        - MCASP Support
    - MCU R5F:
        - UART, SCI Client, IPC, GPIO, I2C, MCAN and MCSPI
        - Example for isolation of MCU through firewall
        - Example for reset isolation of MCU through

+++ source/drivers/bootloader/bootloader_xmodem.c
@@ -96,11 +96,11 @@ int32_t Bootloader_xmodemReceive(uint32_t uartDrvIndex, uint8_t *dstBuf, uint32_
 
 	if(gUartHandle != NULL)
 	{
-		rxBytes = xmodemReceive(dstBuf, maxBufSize);
+		rxBytes = xmodemReceive(dstBuf, (int32_t)maxBufSize);
 
-		if(rxBytes > 0)
+		if(rxBytes > (int32_t)0)
 		{
-			*fileSize = rxBytes;
+			*fileSize = (uint32_t)rxBytes;
 			status = SystemP_SUCCESS;
 		}
 		else
@@ -125,9 +125,9 @@ int32_t Bootloader_xmodemTransmit(uint32_t uartDrvIndex, uint8_t *src, uint32_t 
 
 	if(gUartHandle != NULL)
 	{
-		txBytes = xmodemTransmit(src, len);
+		txBytes = xmodemTransmit(src, (int32_t)len);
 
-		if(txBytes > 0)
+		if(txBytes > (int32_t)0)
 		{
 			status = SystemP_SUCCESS;
 		}
