+++ source/drivers/udma/soc/am62x/udma_soc.c
@@ -200,7 +200,7 @@ void Udma_initDrvHandle(Udma_DrvHandleInt drvHandle)
     drvHandle->iaGemOffset  = CSL_DMSS_GEM_INTA0_SEVI_OFFSET;
     drvHandle->devIdIa      = TISCI_DEV_DMASS0_INTAGGR_0;
     // drvHandle->devIdCore    = Udma_getCoreSciDevId();
-    drvHandle->devIdCore    = Sciclient_getSelfDevIdCore();
+    drvHandle->devIdCore    = (uint16_t)Sciclient_getSelfDevIdCore();
 
 
     /* Init other variables */
