+++ source/kernel/nortos/dpl/common/TimerP.c
@@ -90,7 +90,7 @@ void TimerP_setup(uint32_t baseAddr, TimerP_Params *params)
     DebugP_assert( timerCycles < 0xFFFFFFFFU );
 
     /* calculate count and reload value register value */
-    countVal = 0xFFFFFFFFu - timerCycles - 1U;
+    countVal = 0xFFFFFFFFu - (uint32_t)timerCycles - 1U;
 
     /* keep reload value as 0, later if is auto-reload is enabled, it will be set a value > 0 */
     reloadVal = 0;
